Update patch set 3
Patch Set 3:

(1 comment)

If you check the second patch in the series the first user is there.
The idea is that the user(java code) or the ClusterAdminService would go through there since you also need to register a ShardFrontend into the sharding service from mdsal. The idea is even if the configuration doesn't have a replica on the local member the ShardFrontend is registered into the sharding service so that the requests can be forwarded to the correct shard via the DistributedDataStoreClients.

Patch-set: 3
Label: Verified=0