mdsaltrace utility for debugging
TracingBroker logs 'write' operations and listener registrations to
the md-sal. It logs the instance identifier path,
the objects themselves, as well as the stack trace of the call
invoking the registration or write operation.
It works by operating as a "bump on the stack" between the
application and actual DataBroker, intercepting write
and registration calls and writing to the log.

Change-Id: I107c1f885d3b75d0bb5a43800857630a9a226b03
Signed-off-by: Josh <jhershbe@redhat.com>