Builder of LogicalDatastoreType / InstanceIdentifier / DataObject triple
This standardizes writing well contained code which maps some object
into another one for the data store - of which all "functional" (as in
e.g. netvirt & Co. not "kernel" offset 0) projects are full of - in fact
one could argue OpenDaylight code is mostly just listening to data
changes and then creating new objects... ;-)  

Having a standard Builder interface for this kind of purpose makes
writing such code more obvious, and as folks start adopting this, will
be a better recognizable pattern than the currently widely used private
internal helper methods which receive a lot of arguments spread all over
the place.

This has been initially POC'd in e.g. the
IdentifiedInterfaceWithAclBuilder class of
https://git.opendaylight.org/gerrit/#/c/46206/, where it clearly adds
value for code readability when used in AclServiceTest.  (Note that
while that Gerrit uses this concept for writing tests, the pattern
proposed is more generic and could be used for non-test code as well.)

The proposed use of java.util.Map.Entry<> here isn't because I'm a huge
fan of that API, but because Java doesn't have standard Pair & Tuple
classes (such as e.g. org.eclipse.xtext.xbase.lib.Pair<K, V>), and in
https://git.opendaylight.org/gerrit/#/c/43731/ Robert seemed to advocate
use of Map.Entry. -- Counter proposal on that aspect more than welcome.

Change-Id: I7b1e190e929c74f0692899b336b1b62607ba3d20
Signed-off-by: Michael Vorburger <vorburger@redhat.com>