Update patch set 1
Patch Set 1:

(2 comments)

> Any API changes/improvements should be done in the mdsal project.

Oh, OK.  So where exactly in mdsal should something like this, which is very "close" (related) to e.g. the DataTreeIdentifier which is in the same package as this, go in mdsal - can you point to the correct suitable project? Could anyone else underline this, before I go ahead and move it, Robert, perhaps?

Patch-set: 1