Update patch set 15
Patch Set 15:

(1 comment)

> Patch Set 16: Code-Review+1

Is it good practice to +1 your own patches?  Not saying it's wrong per se but this isn't generally done in the ODL community (at least I don't recall anyone else doing it).  As a committer I actually find it a bit confusing as sometimes I look for another +1 to know it's been reviewed and is ready to merge but then I see it's the patch owner that +1 it.

Patch-set: 15
Label: Verified=0