Update patch set 2
Patch Set 2:

> This is pure sugar, hence should live in an -util package. 

Sure.  Would you +1 this if it was moved to a .util in msdal?

 > As noted
 > in other patches, these APIs are legacy, hence this effort should
 > really be aimed at the mdsal project.
 
I'm unclear how downstream projects which are using the controller and not (yet) the mdsal API would be able to use this then - or am I not getting it, no problem between the two APIs?

 > Furthermore cross-datastore operations happen to work, but we have
 > reached consensus that there is no real use case for them. Hence
 > the LogicalDatastoreType for a particular transaction is an
 > invariant...

Erm... OK.  Interesting.  Now how is this historic API design issue you raise a reason to -1 a proposal for something else unrelated, which happens to have to work with the API as it is today?

Patch-set: 2