Update patch set 2
Patch Set 2: Code-Review-1

Tom I just looked this over (patch set 1) and was about to +1 it, and then saw push a patch set 2 which mixes in non-CS related real code changes, was that intentional, or perhaps a mistake? I'm +1 the CS related clean up, but don't understand enough to be able to judge the patch set 2 changes.  IMHO it's best to submit such things separately?

Patch-set: 2