Update patch set 6
Patch Set 6:

> I'm not that familiar with the various pom parents but I trust you did the right thing :)

To be honest, it's not 100% perfect as I could have made a more accurate use of the parents (for instance don't use config-parent but bundle-parent, they are a few occurrence like this in this patch). But the real intention is to start getting rid of opendaylight/commons/opendaylight/pom.xml

So I will have to amend a new patch set to address this. But I wanted to have both Stephen and your opinion before moving forward.

Patch-set: 6
Label: Verified=0