Update patch set 2
Patch Set 2:

Those are two separate concerns:

Pure sugar is about API design itself, e.g. adding a slew of default methods which are just wrappers on top of existing interfaces.

The comment about invariants goes into what we want to be promising and can provide as a compatibility layer. Since the store type is an invariant, it should be treated as such from API design perspective -- as a user I should not be specifying those over and over again.

Patch-set: 2