Update patch set 1
Patch Set 1:

> Please do no use commons/opendaylight/pom.xml - trying to get away
 > from it: https://git.opendaylight.org/gerrit/#/q/topic:remove-commons-opendaylight

Alexis, thanks for your feedback, but ... I'm not following the logic you apply here, I'm afraid.  This change is part of the overall effort described here: https://lists.opendaylight.org/pipermail/odlparent-dev/2016-October/000565.html.  This commons/opendaylight/pom.xml exists today.  You cannot block (-1) a minor clean up effort, this change, by saying that something "shouldn't be used" - it is!  And if it's not used, then please remove this, and I'll abandon this change.  As long as it is there, why block this clean-up??

Patch-set: 1