Fix PMD warnings in sal-akk-raft and enable enforcement
There was only 1 warning for duplicated classes:
org.opendaylight.controller.cluster.raft.persisted.ServerInfo and
org.opendaylight.controller.cluster.raft.ServerConfigurationPayload.ServerInfo.
The former is an intentional copy of the latter and the latter is deprecated
and will be removed at some point. To workaround the warning I changed a
couple lines in the new version so it differs.

Change-Id: I1cec0b48c677f792b65800926729fabd0a3cf61c
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>