Update patch set 4
Patch Set 4:

Note the SchemaContextSingleton intro. in Patch Set 4 .. on self reviewing this, I realized that the original solution would only have cached the expensive SchemaContext per *Test class, which was good and progress compared to once per @Test as originally, but only a first step.. it would still have re-created it after FirstTest in a SecondTest class. With the SchemaContextSingleton, it will truly only be created once, and shared among a FirstTest and a SecondTest class.

Patch-set: 4
Label: Verified=0