Update patch set 2
Patch Set 2:

(1 comment)

FYI: In all new tests I write I'm only using the DataBrokerTestModule anyway, which this changes to use the ConstantSchemaAbstractDataBrokerTest instead of the AbstractDataBrokerTest, never directly the AbstractDataBrokerTest or ConstantSchemaAbstractDataBrokerTest - this is more of a convenience for anyone else who likes to do extends of infra super classes in tests (I don't).  

To be honest, I was initially tempted to even just change the AbstractDataBrokerTest directly, but then thought I would get push back because strictly speaking this changes behaviour, so for backward compatibility I've intro. the new class.. but not sure how relevant this change really is in practice - like do you think anyone could have an existing test extending AbstractDataBrokerTest which would break by this optimiation? If you don't see that as a problem either, then I would be happy to just fold this into the AbstractDataBrokerTest - and give everyone using it faster tests out-of-the-box, without them having to change from AbstractDataBrokerTest to ConstantSchemaAbstractDataBrokerTest. For DataBrokerTestModule, which admittedly is much more recent and only just starting to be used, I'm doing that anyway already here.

Patch-set: 2