Update patch set 6
Patch Set 6:

I recall now the intent of the original implementation, see https://wiki.opendaylight.org/view/Using_Blueprint#Using_property-placeholder, specifically "It is recommended to use restart-dependents-on-updates in lieu of the reload update-strategy".  So it is intended for update-strategy to be set to none if restart-dependents-on-updates is set. restart-dependents-on-updates is set on the blueprint element b/c we can't extend property-placeholder. We also can't extend the update-strategy enum. 

Given that, I think what we should do, if anything, is only honor restart-dependents-on-updates if update-strategy is set to none so we don't reload the container twice.

Patch-set: 6