Update patch set 6
Patch Set 6:

> > > Given that, I think what we should do, if anything, is only honor
restart-dependents-on-updates if update-strategy is set to none so
we don't reload the container twice.

> > I’m not sure to fully get that statement, as if I have both
restart-dependents-on-updates = true and update-strategy = none
it’s reloading only once. Same if I have restart-dependents-on-updates
= true and update-strategy = reload. (this is true at least with
https://git.opendaylight.org/gerrit/#/c/48284/)

> I was referring to restart-dependents-on-updates=true and update-strategy=reload. I would expect both our code and the Aries code to independently restart the container. If you tried that and see only one restart then maybe one of the ManagedService OSGi registrations gets dropped if there's already one registered. Regardless, you wouldn't want both enabled. restart-dependents-on-updates is intended to replace the behavior of update-strategy=reload.

I'll double check tomorrow. Because I recall seeing two ManagedService services while executing the "services" command in the karaf cli for the topoprocessing bundle, which lead me to some digging but then got side-track and forget and haven't looked back.

Patch-set: 6