Update patch set 2
Patch Set 2:

> This will leak the contents, as iterator will go through, but will
 > clear also the exception.
 > 
 > Can we add some javadocs around protected methods about their
 > interactions and some logs to know what is going on?

I agree it could leak contents. This is an edge case where the container is destroyed immediately after and while it's starting up. This isn't likely to happen in production but can happen during feature tests. I had assumed  the container would provide the protection but apparently it doesn't. We should explicitly synchronize the list for atomicity. Also prior to adding I think we should check if it's being destroyed (add a flag).

Patch-set: 2