Update patch set 2
Patch Set 2:

> I am not sure about this one. I think we really want to remove the
 > optimization and replicate even empty transactions -- it will be
 > required for BUG-5280 so followers have an accurate view of
 > transaction states. I do not think persisting an empty entry will
 > hurt that much, especially with pipelining in place and has the
 > benefit of keeping the code more straightforward...

I tend to agree with this and I considered removing the empty candidate check. In fact I'd like to remove the "canSkipCandidate" path as well to keep it simple. For now though let's keep it as is and merge this to fix the immediate problem and unblock folks. I'll do a follow-up where I can take my time (I'm actually on PTO until next Tue) and add appropriate unit test cases.

Patch-set: 2
Label: Verified=0