Bug 7521: Convert Snapshot to store a ByteSource
Created a new Snapshot class that stores the state as a ByteSource
instead of a byte[] to lay the groundwork for handling snapshots > 2G.
The previous Snapshot class was deprecated and "read resolves" to the
new class on de-serialization.

Currently the ByteSource instances passed to Snapshot are still backed
by byte[] but will be changed in a subsequent patch.

Change-Id: I2ea30870b54478d7ef5669335ca4b444539f8d56
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>