Update patch set 6
Patch Set 6:

> > I'm not a fan of disabling the transitive dependency analysis
 > 
 > Is there a readable documentation on what exactly is
 > includeTransitiveDependency supposed to do?

Not as far as I know. There's some info on https://issues.apache.org/jira/browse/KARAF-931 which shows that the option was added precisely to allow the kind of processing Robert wanted in this patch!

Basically, the way it works is as follows. The Karaf Maven plugin can generate a feature descriptor based on the corresponding POM; by default, the generated feature will only include features and bundles explicitly listed in the POM. Activating includeTransitiveDependency means that the feature will also include the bundles that are transitively depended on; in Karaf 4.0.9 (and 4.0.8-odl), these transitive dependencies are ignored if they're provided by another feature the feature depends on (see https://issues.apache.org/jira/browse/KARAF-884 for details).

 > I am fan of skipping the analysis and letting SFT (in verify and in
 > distribution-check) decide whether the feature is instalable
 > enough.

I think there are two ways of considering things.

* We can require developers to list all the bundles and features their features need to depend on, including any required transitive dependencies. This is the current situation, and it works up to a point, but becomes painful especially down in offset-2 land: whenever a third-party dependency changes (e.g. thanks to odlparent), some bundles may need to be added; and there's a widespread practice of using ODL offset-0 and offset-1 bundles instead of features, with a random selection of bundles required to get things to work (and usually, this is determined by a process of adding things until the feature works, and then leaving everything and never cleaning anything up because no one understands what's going on). So when you look at a feature definition, you never know whether a bundle is included because it's part of the feature's specification, or if it's just there because it's needed to make things work (or used to be needed, or was thought to be needed).

* We can allow developers to only list the bundles they actually *care* about exporting in their features, i.e. those which provide the API they want to support, and implement whatever the feature is supposed to do. Then the plugin handles all the other bundles they need.

In both cases SFT is very useful. I much prefer the second approach though because it results in feature POMs which describe what the developers *mean*, not what they need to specify because of bundle requirements. It goes along the lines of specifying all the dependencies a bundle depends on (including those pulled in transitively), but ignoring everything else — we shouldn't need to change bundle definitions because some upstream changed, the build system should be able to figure things out on its own.

Currently the situation is somewhat imperfect because the Maven plugin doesn't know about everything our distribution includes, but I'm working on improving that. If you want to see a great example of the benefits using the plugin can provide, check out ttp's generated features...

Patch-set: 6