odl-config-persister feature clean-up
This removes all third-party bundles which are pulled in
transitively. The resulting feature file is identical, but the POM
only declares the ODL bundles we really care about, and
org.apache.karaf.util which is needed for some reason (and is wrapped,
so it's not provided by our distribution).

This is one possible variant of three (see the other patches in the
same topic).

Change-Id: Ia5a452bb1fd352955a3c585a9b98b344c8df7914
Signed-off-by: Stephen Kitt <skitt@redhat.com>