Update patch set 1
Patch Set 1:

> Tom, Robert: *ALL* of your review comments pertains to code that
 > has been like this for years - I've merely refactored things
 > here... in general, I don't think we can fix the world in every
 > Gerrit that touches something?  However, in order to move forward
 > with this and get this finally merged (people need this; note e.g.
 > https://git.opendaylight.org/gerrit/#/c/51577/ work around in
 > Genius), I'll try to apply your feedback to this existing code
 > anyway...

I understand you've moved the code from elsewhere but the changes are minor and presents a chance to improve it a bit. Increasing the timeout could actually be significant as before with the direct executor the future was always already complete. Now a separate thread in the mix introduces context switching and, even though 500 ms is a relatively long time, it could trip if conditions on a particular run are right. I think it's better to be safe with a larger cushion.

Patch-set: 1