Update patch set 2
Patch Set 2:

> > I understand you've moved the code from elsewhere but the changes
 > > are minor and presents a chance to improve it a bit.
 > 
 > Oh right, now I get what you were after here... hope better in
 > current patch set?
 > 
 > > Increasing the
 > > timeout could actually be significant as before with the direct
 > > executor the future was always already complete. Now a separate
 > > thread in the mix introduces context switching and, even though
 > 500
 > > ms is a relatively long time, it could trip if conditions on a
 > > particular run are right. I think it's better to be safe with a
 > > larger cushion.
 > 
 > Personally I've just used get() without timeout to wait for
 > DataBroker commits in tests... I understand it could theoretically
 > wait "forever", but... only if there was really bad implementation
 > bug, no?  And surefire would eventually kill it anyway?

I don't know if surefire would kill it but I've seen tests take hours on jenkins if you let it :)  I always use timeouts (5 sec is my standard) in such cases as get() and await(). I've seen tests hang on a get(). Granted it shouldn't happen but that's why we write unit tests to verify :)  It's a little annoying in eclipse b/c nothing will kill it and then you have to run with the debugger and suspend to find where it's hanging. Adding generous timeouts saves you the trouble.

Patch-set: 2